Split byte buffer tests
